+++ website/events/templates/mails/invoice_just_created_subject.txt
@@ -1,1 +1,1 @@
-Nuevo patrocinio creado
+Nuevo factura asoiciada a patroncinio
